Geocoder as an interface to allow for chained and cached types
